+++ scripts/utils/migrate/converters/to-markdown.js
@@ -1,5 +1,5 @@
 const htmlToMarkdown = require('../html-to-markdown');
-const getFrontmatter = require('../../../frontmatter/get-frontmatter');
+const getFrontmatter = require('../get-frontmatter');
 
 const toMarkdown = (file) => {
   file.contents = getFrontmatter(file.meta.doc) + htmlToMarkdown(file.contents);
